+++ samples/csharp/getting-started/TimeSeries_PowerAnomalyDetection/PowerAnomalyDetection/DataStructures/SpikePrediction.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Microsoft.ML.Data;
+
+namespace PowerAnomalyDetection.DataStructures
+{
+    class SpikePrediction
+    {
+        [VectorType(3)]
+        public double[] Prediction { get; set; }
+    }
+}
